requirements.txt
agate==1.6.3
attrs==23.1.0
Babel==2.12.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.4
dbt-core==1.1.5
dbt-extractor==0.4.1
dbt-mysql==1.1.0
dnspython==2.4.1
future==0.18.3
hologram==0.0.15
idna==3.4
isodate==0.6.1
Jinja2==2.11.3
jsonschema==3.2.0
jsonschema-specifications==2023.7.1
leather==0.3.4
Logbook==1.5.3
MarkupSafe==2.0.1
mashumaro==2.9
minimal-snowplow-tracker==0.0.2
msgpack==1.0.5
mysql-connector-python==8.0.33
networkx==2.8.3
packaging==21.3
parsedatetime==2.4
pathspec==0.11.2
protobuf==3.20.3
pycparser==2.21
pymongo==4.4.1
pyparsing==3.1.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-slugify==8.0.1
pytimeparse==1.1.8
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.0
requests==2.31.0
rpds-py==0.9.2
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.1.2